# main-webbase/docs/swagger.yaml
basePath: /
definitions:
  controllers.FeedCursorEnvelope:
    properties:
      has_more:
        example: true
        type: boolean
      items:
        items:
          additionalProperties: {}
          type: object
        type: array
      next_cursor:
        example: eyJpZCI6IjY4ZTYzMDAyZGYyMjVjZDk1NTE3M2RiIn0=
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid body
        type: string
    type: object
...
  dto.PostAs:
    properties:
      label:
        description: e.g., "Head • SMO"
        type: string
      org_path:
        type: string
      position_key:
        type: string
    type: object
  dto.PostResponse:
    properties:
      categoryIds:
        example:
        - '[''68bd8d30b98a8dce0eab0db6'''
        - '''68bd8d30b98a8dce0eab0db7'']'
        items:
          type: string
        type: array
      commentCount:
        example: 0
        type: integer
      createdAt:
        example: "2025-09-07T13:47:47Z"
        type: string
      hashtag:
        items:
          type: string
        type: array
      is_liked:
        example: false
        type: boolean
      likeCount:
        example: 0
        type: integer
      media:
        example:
        - '[''ttp://45.144.166.252:46602/uploads/cat.png'']'
        items:
          type: string
        type: array
      name:
        example: JY
        type: string
      org_of_content:
        description: this content belongs to which org (org_path)
        type: string
      postAs:
        allOf:
        - $ref: '#/definitions/dto.PostAs'
        description: which role is posting this
      postText:
        example: สวัสดี KU!
        type: string
      status:
        example: active
        type: string
      updatedAt:
        example: "2025-09-07T13:47:47Z"
        type: string
      userId:
        example: 66c6248b98c56c39f018e7d2
        type: string
      username:
        example: jy_smo
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/dto.Visibility'
        description: which roles can see this post, array of roleIds
    type: object
...
  dto.Visibility:
    properties:
      access:
        type: string
      audience:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a sample server for user management.
  title: Fiber MongoDB API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Return all unread notifications and the total count for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: Unread notification count and list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch notifications
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List unread notifications for the current user
      tags:
      - notifications
  /notifications/{id}:
    get:
      description: Fetch a notification by ID for the authenticated user, mark it
        as read, and return the updated document.
      parameters:
      - description: Notification ID (hex ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated notification document
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Notification not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update notification
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a notification and mark it as read
      tags:
      - notifications
...
  /posts:
    get:
      description: List posts that the viewer has permission to see, based on their
        organizational unit and position.
      parameters:
      - default: 10
        description: Max items per page
        in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      - description: Opaque next-page cursor (base64)
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.FeedCursorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get posts visible to the viewer
      tags:
      - feed
    post:
      consumes:
      - multipart/form-data
      description: Create a new post; supports multipart form upload
      parameters:
      - description: Upload media file
        in: formData
        name: file
        type: file
      - description: Post text
        in: formData
        name: postText
        required: true
        type: string
      - description: Organization of content
        in: formData
        name: org_of_content
        type: string
      - description: Organization path
        in: formData
        name: postAs.org_path
        required: true
        type: string
      - description: Position key
        in: formData
        name: postAs.position_key
        required: true
        type: string
      - description: 'Visibility (default: public)'
        in: formData
        name: visibility.access
        type: string
      - description: Category IDs (repeatable)
        in: formData
        name: categoryIds
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a post with optional media upload
      tags:
      - posts
...
swagger: "2.0"